refactor(cli): migrate info command tests to colocated structure
Move info command tests from src/__tests__/info-command.test.ts to
commands/info/__tests__/index.test.ts following CLI testing patterns.

Changes:
- Import infoCommand directly instead of program
- Organize tests under "system information display" describe block
- Add explicit API host assertion
